+++ financius/src/main/java/com/code44/finance/ui/transactions/list/TransactionsAdapter.java
@@ -1,5 +1,6 @@
 package com.code44.finance.ui.transactions.list;
 
+import android.content.Context;
 import android.database.Cursor;
 import android.support.v7.widget.RecyclerView;
 import android.text.SpannableStringBuilder;
@@ -21,6 +22,7 @@ import com.code44.finance.money.AmountFormatter;
 import com.code44.finance.ui.common.adapters.ModelsAdapter;
 import com.code44.finance.ui.common.presenters.ModelsActivityPresenter;
 import com.code44.finance.ui.common.recycler.SectionsDecoration;
+import com.code44.finance.utils.AccountUtils;
 import com.code44.finance.utils.CategoryUtils;
 import com.code44.finance.utils.TextBackgroundSpan;
 import com.code44.finance.utils.ThemeUtils;
@@ -77,9 +79,6 @@ class TransactionsAdapter extends ModelsAdapter<Transaction> implements Sections
     private static class ViewHolder extends ModelViewHolder<Transaction> {
         private final AmountFormatter amountFormatter;
         private final int textColorPrimary;
-        private final String unknownExpenseTitle;
-        private final String unknownIncomeTitle;
-        private final String unknownTransferTitle;
         private final int tagBackgroundColor;
         private final float tagBackgroundRadius;
         private final int expenseAmountColor;
@@ -87,22 +86,19 @@ class TransactionsAdapter extends ModelsAdapter<Transaction> implements Sections
         private final int transferAmountColor;
         private final int weakColor;
 
-        private final ImageView color_IV;
-        private final TextView weekday_TV;
-        private final TextView day_TV;
-        private final TextView title_TV;
-        private final TextView subtitle_TV;
-        private final TextView amount_TV;
-        private final TextView account_TV;
+        private final ImageView colorImageView;
+        private final TextView weekdayTextView;
+        private final TextView dayTextView;
+        private final TextView titleTextView;
+        private final TextView subtitleTextView;
+        private final TextView amountTextView;
+        private final TextView accountTextView;
 
         public ViewHolder(View itemView, AmountFormatter amountFormatter) {
             super(itemView);
             this.amountFormatter = amountFormatter;
 
             textColorPrimary = ThemeUtils.getColor(itemView.getContext(), android.R.attr.textColorPrimary);
-            unknownExpenseTitle = itemView.getContext().getString(R.string.expense);
-            unknownIncomeTitle = itemView.getContext().getString(R.string.income);
-            unknownTransferTitle = itemView.getContext().getString(R.string.transfer);
             tagBackgroundColor = ThemeUtils.getColor(itemView.getContext(), R.attr.backgroundColorSecondary);
             tagBackgroundRadius = itemView.getContext().getResources().getDimension(R.dimen.tag_radius);
             expenseAmountColor = ThemeUtils.getColor(itemView.getContext(), android.R.attr.textColorPrimary);
@@ -110,26 +106,26 @@ class TransactionsAdapter extends ModelsAdapter<Transaction> implements Sections
             transferAmountColor = ThemeUtils.getColor(itemView.getContext(), R.attr.textColorNeutral);
             weakColor = ThemeUtils.getColor(itemView.getContext(), android.R.attr.textColorSecondary);
 
-            color_IV = (ImageView) itemView.findViewById(R.id.colorImageView);
-            weekday_TV = (TextView) itemView.findViewById(R.id.weekdayTextView);
-            day_TV = (TextView) itemView.findViewById(R.id.dayTextView);
-            title_TV = (TextView) itemView.findViewById(R.id.titleTextView);
-            subtitle_TV = (TextView) itemView.findViewById(R.id.subtitleTextView);
-            amount_TV = (TextView) itemView.findViewById(R.id.amountTextView);
-            account_TV = (TextView) itemView.findViewById(R.id.accountTextView);
+            colorImageView = (ImageView) itemView.findViewById(R.id.colorImageView);
+            weekdayTextView = (TextView) itemView.findViewById(R.id.weekdayTextView);
+            dayTextView = (TextView) itemView.findViewById(R.id.dayTextView);
+            titleTextView = (TextView) itemView.findViewById(R.id.titleTextView);
+            subtitleTextView = (TextView) itemView.findViewById(R.id.subtitleTextView);
+            amountTextView = (TextView) itemView.findViewById(R.id.amountTextView);
+            accountTextView = (TextView) itemView.findViewById(R.id.accountTextView);
         }
 
         @Override protected void bind(Transaction transaction, Cursor cursor, int position, ModelsActivityPresenter.Mode mode, boolean isSelected) {
             final DateTime date = new DateTime(transaction.getDate());
 
             // Set values
-            weekday_TV.setText(date.dayOfWeek().getAsShortText());
-            day_TV.setText(date.dayOfMonth().getAsShortText());
-            color_IV.setColorFilter(CategoryUtils.getColor(itemView.getContext(), transaction));
-            amount_TV.setText(amountFormatter.format(transaction));
-            title_TV.setTextColor(textColorPrimary);
-            title_TV.setText(getTitle(transaction));
-            subtitle_TV.setText(getSubtitle(transaction));
+            weekdayTextView.setText(date.dayOfWeek().getAsShortText());
+            dayTextView.setText(date.dayOfMonth().getAsShortText());
+            colorImageView.setColorFilter(CategoryUtils.getColor(itemView.getContext(), transaction));
+            amountTextView.setText(amountFormatter.format(transaction));
+            titleTextView.setTextColor(textColorPrimary);
+            titleTextView.setText(getTitle(titleTextView.getContext(), transaction));
+            subtitleTextView.setText(getSubtitle(transaction));
             bindViewForTransactionType(transaction);
 
             if (transaction.getTransactionState() == TransactionState.Pending) {
@@ -138,27 +134,24 @@ class TransactionsAdapter extends ModelsAdapter<Transaction> implements Sections
         }
 
         private void bindViewForTransactionType(Transaction transaction) {
-            final String account;
+            final String account = AccountUtils.getTitle(transaction);
             final int amountColor;
             switch (transaction.getTransactionType()) {
                 case Expense:
-                    account = transaction.getAccountFrom() != null ? transaction.getAccountFrom().getTitle() : UNKNOWN_VALUE;
                     amountColor = expenseAmountColor;
                     break;
                 case Income:
-                    account = transaction.getAccountTo() != null ? transaction.getAccountTo().getTitle() : UNKNOWN_VALUE;
                     amountColor = incomeAmountColor;
                     break;
                 case Transfer:
-                    account = (transaction.getAccountFrom() != null ? transaction.getAccountFrom().getTitle() : UNKNOWN_VALUE) + TRANSFER_SYMBOL + (transaction.getAccountTo() != null ? transaction.getAccountTo().getTitle() : UNKNOWN_VALUE);
                     amountColor = transferAmountColor;
                     break;
                 default:
                     throw new IllegalArgumentException("Transaction type " + transaction.getTransactionType() + " is not supported.");
             }
 
-            account_TV.setText(account);
-            amount_TV.setTextColor(amountColor);
+            accountTextView.setText(account);
+            amountTextView.setTextColor(amountColor);
         }
 
         private void bindViewPending(Transaction transaction) {
@@ -166,12 +159,12 @@ class TransactionsAdapter extends ModelsAdapter<Transaction> implements Sections
             final boolean isTransfer = transaction.getTransactionType() == TransactionType.Transfer;
 
             if (isCategoryUnknown && !isTransfer) {
-                title_TV.setTextColor(weakColor);
-                color_IV.setColorFilter(weakColor);
+                titleTextView.setTextColor(weakColor);
+                colorImageView.setColorFilter(weakColor);
             }
 
             if (transaction.getAmount() == 0) {
-                amount_TV.setTextColor(weakColor);
+                amountTextView.setTextColor(weakColor);
             }
 
             final boolean isAccountFromUnknown = transaction.getAccountFrom() == null || !transaction.getAccountFrom().hasId();
@@ -180,46 +173,33 @@ class TransactionsAdapter extends ModelsAdapter<Transaction> implements Sections
             final boolean isIncome = transaction.getTransactionType() == TransactionType.Income;
             if (isExpense) {
                 if (isAccountFromUnknown) {
-                    account_TV.setText(UNKNOWN_VALUE);
+                    accountTextView.setText(UNKNOWN_VALUE);
                 }
 
                 if (transaction.getAmount() > 0) {
-                    amount_TV.setTextColor(expenseAmountColor);
+                    amountTextView.setTextColor(expenseAmountColor);
                 }
             } else if (isIncome) {
                 if (isAccountToUnknown) {
-                    account_TV.setText(UNKNOWN_VALUE);
+                    accountTextView.setText(UNKNOWN_VALUE);
                 }
 
                 if (transaction.getAmount() > 0) {
-                    amount_TV.setTextColor(incomeAmountColor);
+                    amountTextView.setTextColor(incomeAmountColor);
                 }
             } else {
                 final String accountFrom = isAccountFromUnknown ? UNKNOWN_VALUE : transaction.getAccountFrom().getTitle();
                 final String accountTo = isAccountToUnknown ? UNKNOWN_VALUE : transaction.getAccountTo().getTitle();
-                account_TV.setText(accountFrom + TRANSFER_SYMBOL + accountTo);
+                accountTextView.setText(accountFrom + TRANSFER_SYMBOL + accountTo);
                 if (transaction.getAmount() > 0) {
-                    amount_TV.setTextColor(transferAmountColor);
+                    amountTextView.setTextColor(transferAmountColor);
                 }
             }
         }
 
-        private String getTitle(Transaction transaction) {
+        private String getTitle(Context context, Transaction transaction) {
             if (Strings.isEmpty(transaction.getNote())) {
-                if (transaction.getCategory() != null && transaction.getCategory().hasId()) {
-                    return transaction.getCategory().getTitle();
-                } else {
-                    switch (transaction.getTransactionType()) {
-                        case Expense:
-                            return unknownExpenseTitle;
-                        case Income:
-                            return unknownIncomeTitle;
-                        case Transfer:
-                            return unknownTransferTitle;
-                        default:
-                            throw new IllegalArgumentException("Transaction type " + transaction.getTransactionType() + " is not supported.");
-                    }
-                }
+                return CategoryUtils.getTitle(context, transaction);
             } else {
                 return transaction.getNote();
             }